requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
networkx>=3.1
matplotlib>=3.7.0
scipy>=1.10.0
scikit-learn>=1.2.0
spacy>=3.5.0

# API and async
aiohttp>=3.8.4
tenacity>=8.2.0

# NLP and embeddings
sentence-transformers>=2.2.2

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# For visualization (future use)
# gradio>=3.50.0
# streamlit>=1.28.0

# For testing
pytest>=7.4.0
pytest-asyncio>=0.21.1